Use clock by default (instead of clock.check)
We're seeing clocks desynchronized.  My theory is that this might
be because ntp can take five minutes or more to actually sync the
clocks, and clock.check doesn't do any setting of the clocks, just
reporting.  clock, OTOH, stops ntpd and does an ntpdate, and then
restarts ntpd, which should kickstart it with a much-closer-to-correct
time.

Signed-off-by: Dan Mick <dan.mick@redhat.com>