Ceph-disk to use correct user in check_journal_req
The ceph-disk tool calls ceph-osd to check the journal requirements
using OSD id 0. This creates a log file for osd-0 on the system
using the current user/group for file permissions.  When ceph-disk
is run as root this makes the file owned by root which makes
the osd daemon for osd.0 unable to write to its own log file.

This commit changes the journal reqs calls of ceph-osd to pass
the ceph user and group so ceph-osd creates the log file with the
appropriate permissions.

Fixes: http://tracker.ceph.com/issues/18538

Signed-off-by: Samuel Matzek <smatzek@us.ibm.com>
(cherry picked from commit bcf7514bf53693ec61e482341787c80494589faf)