rgw: add rgwrados init_barrier() (PARTIAL, PROPOSAL)
To avoid faults in threads spawned by RGWRados::initialize() and
it's friend init_complete(), force the thread instances to execute
a barrier on RGWRados that can prevent any further accesses.

An assumption here is that if the integration for simpler cases
such as RGWGC and RGWObjectExpirer looks good, then it will be only
a bit more work to deal with cases such as WorkQueue descendants
(which currently lack a place to hook RGWRados::init_barrier()).

Another assumption is that the slow path is the main path, which
includes the call that completes all barriers.  On this assumption,
we can avoid the requirement to track and complete waits by callers
of RGWRados::init_barrier(), and just wait for a short period if
initialization has failed.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>