test: use the test name as the default prefix of temp pool name
the prefix was "test-rados-api-" if not specified, but it does not help
much if multiple tests are running in parallel and all of them are using
the same prefix, also the PID used to compose the pool name is not the
one tracked by workunits/rados/test.sh. so we need something more
discriminative to name the temp pool for testing.

Signed-off-by: Kefu Chai <kchai@redhat.com>