osdc/Objecter: refactor list_nobjects to use hobject_t as position
Stop using current_pg as a position pointer; use the hobject_t
cursor explicitly.

We keep current_pg *only* for compatibility with !sortbitwise
clusters, and we only use it when we get back MAX from a
!sortbitwise OSD and need to determine where the start of the next
PG is.  In !sortbitwise mode we also have the legacy kludges to
behave on PG split.

Signed-off-by: Sage Weil <sage@redhat.com>