tests: osd-scrub-repair.sh disable scrub backoff in test
there is good chance that the newly scheduled scrub fail to pass the
flip-coin check: the probability of scheduling a new scrub 1/3 every
time we flip the coin. so, instead hardwiring the probability to 1/3,
a new option named "osd_scrub_backoff_ratio" is introduced. so we can
override it when performing test.

Signed-off-by: Kefu Chai <kchai@redhat.com>