osd: drop unused+inefficient support for pg[n]ls at a specific snapid
This was super slow, and Objecter was incapable of generating the requests
to use it.

To do this properly we should create a new listing op that returns the
set of clones and/or snaps for each object as part of a single listing
result.  If/when the need arises.

Signed-off-by: Sage Weil <sage@redhat.com>