rgw: make sending Content-Length in 204 and 304 controllable
This commit introduces a new configurable "rgw print prohibited
content length" to let operator decide whether RadosGW complies
to RFC 7230 (a part of the HTTP specification) or violates it
but follows the Swift's behavior.

This commit has not been backported from master but instead
made specifically for Jewel because of significant differences
between it and master in the front-end architecture. However,
the meaning and logic of the newly introduced configurable has
been preserved.

Fixes: http://tracker.ceph.com/issues/16602
Signed-off-by: Radoslaw Zarzynski <rzarzynski@mirantis.com>