ceph-disk: dmcrypt activate must use the same cluster as prepare
When dmcrypt is used, the fsid cannot be retrieved from the data
partition because it is encrypted. Store the fsid in the lockbox to
enable dmcrypt activation using the same logic as regular activation.

The fsid is used to retrive the cluster name that was used during
prepare, reason why activation does not and must not have a --cluster
argument.

Fixes: http://tracker.ceph.com/issues/17821

Signed-off-by: Loic Dachary <ldachary@redhat.com>
(cherry picked from commit 7f66672b675abbc0262769d32a38112c781fefac)

Conflicts:
	src/ceph-disk/ceph_disk/main.py