Revert "Revert "ceph-disk: change get_dmcrypt_key test to support different cluster name""
This reverts commit 105939b16f71640a0e7c924c80f247542fe86137 which
reverted 2082f636c6781fcd614eb78417ed8d419ba378e5. Although in the same
pull request, 2082f636c6781fcd614eb78417ed8d419ba378e5 was a fix for
1e0017123359fe4c0e9eaea89e4024017e6926ae. The two should have been
squashed but they were not.

When reverting the pull request I reverted all the commits in the pull
request but I did it in the wrong order and the fix became the
problem. Luckily the outcome is a noop instead of a bug, reason why
it was undetected by the tests.

Signed-off-by: Loic Dachary <ldachary@redhat.com>