rgw/openssl fix: xenial autoconf logic problem: gcc/ld got too smart...
On xenial, cc -o foo -lssl -lcrypto doesn't always
record libssl.so libcrypto.so as runtime library dependencies.
It is necessary to actually *use* a function from the library before
it gets recorded.  The ld(1) options "--as-needed" and "no-as-needed"
control this.  Evidently the default has changed in xenial.
That caused my smart "soname" detecting logic for openssl to
stop working.  To make it work, the test program has to
actually reference routines or variables inside the library.

This is a quick fix for xenial / autoconf.  There needs to be
a better fix for cmake and master.

Signed-off-by: Marcus Watts <mwatts@redhat.com>