ReplicatedBackend: clear pull source once we are done with it
f0f813d857282f964c1a6a4fa6f36d3124748cf9 factored out the clear_pull
behavoir, but we actually need to clear the pull source in
handle_pull_response (even though we don't clear the pulling entry
until the callback) so that we don't clear the pulling entry in
check_recovery_sources.  This should restore the clearing behavior
to basically what it was in kraken.

Fixes: http://tracker.ceph.com/issues/19076
Signed-off-by: Samuel Just <sjust@redhat.com>