mgr: improve handle_command
Run the python module calls in a finisher so that
they don't block the daemonserver lock and so that
they can call back into mgr stuff if they need to.

Fix passing through commands to python modules, this
was giving EINVAL because only things with a MgrCommand
were getting let in.

Also fix get_command_descriptions, which was not
including the output of the formatter in the response.

Signed-off-by: John Spray <john.spray@redhat.com>