osd/ReplicatedPG: try with pool's use-gmt setting if hitset archive not found
due to http://tracker.ceph.com/issues/19185, a hammer OSD could store
pg_hit_set_info_t with "use-gmt = true" even the pool setting is false.
so we use the pool setting as a fallback if the hitset archive is not
found locally and the pool.use_gmt does not match with hitset.use_gmt.

Fixes: http://tracker.ceph.com/issues/19185
Signed-off-by: Kefu Chai <kchai@redhat.com>

Conflicts:
  osd/ReplicatedPG.cc:  This cannot be cherry-picked from master
    because: hammer should upgrade to jewel first before moving to a
    higher version. so there is no necessary to include this workaround
    in releases later than jewel.