include/denc: remove nullptr runtime magic boundedness check
We were passing *(const T*)nullptr to bound_encode so that we
would segv at runtime in a bound_encode implementation had anyd
dependency on the value.  However, this relies on undefined
language behavior and fails on low optimization levels.

So drop these checks and rely on the developer to implement a
legal bound_encode()

Signed-off-by: Sage Weil <sage@redhat.com>