rgw: add a configurable for enforcing swift bucket name validation
Swift containers are declared in the upstream documentation as 1-256
bytes of UTF-8, and the only forbidden character is '/'.

We did not previously enforce the no-slash rule.

Adding a configurable `rgw_swift_enforce_bucket_names` which validates
according to swift container naming restrictions (which currently
forbids "/" in container names), also dropping the 0xff check that we
previously had since check_utf8 should already catch that

Making this as a configurable with the default that we enforce the swift
naming rules of no slash, the configurable is useful in cases where
when previously created buckets with slashes exists and need to be used

See-Also: https://docs.openstack.org/developer/swift/api/object_api_v1_overview.html
Fixes: http://tracker.ceph.com/issues/19264

Signed-off-by: Robin H. Johnson <robin.johnson@dreamhost.com>
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>