FileStore::statfs() avoid double-counting the journal disk space
If the journal is an ordinary file and is located on the same
filesystem as the data FileStore::statfs() double counts the space
occupied by the journal: statfs (the system call) already takes
into account the space allocated for the journal file. Therefore
the OSD might be prematurely marked as full (near full), especially
if the data drive (partition) is relatively small (say, an SSD).

Signed-off-by: Alexey Sheplyakov <asheplyakov@mirantis.com>