mds: set ceph-mds name uncond for external tools
External tools like logrotate which use killall rely on the "ceph-mds" name to
find it in /proc/*/comm. This is normally the case but when ceph-mds respawns
using /proc/self/exe (on Linux), its name will change to "exe". This makes
logrotate fail to signal ceph-mds the log needs to be reopened and will lead to
the log growing until the disk is full.

This patch unconditionally sets the name so on respawn the name is correct (and
in any other possible scenario, such as a misnamed executable).  Note, there is
still a very small race between execve and ceph_pthread_setname where the
process name is wrong.

Problem was introduced by aa3fd39fa9e60108c3023a53232df9cbb72f667a.

Fixes: http://tracker.ceph.com/issues/19291

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>