jewel: osd/PGLog: reindex properly on pg log split
When PGLog::IndexedLog::split_into runs it builds the list, which means
the old indices are wrong (point to bad memory), but index() will not
rebuild them because ever since b858e86 we won't rebuild them if they
are already built. Fix that by calling unindex() before the split.

Based on 643ae42cf27f16dd6ed4e1402acc0483bb9fca74. Notice that both
the child and the parent log are re-indexed in Jewel, so the only
problem is missing unindex().

Signed-off-by: Alexey Sheplyakov <asheplyakov@mirantis.com>