tests: ceph_test_rados_api_watch_notify: test timeout using rados_watch3()
objecter resends the watch request upon tcp reconnecting, and OSD
resets the watch timeout when handling the resent watch request, so,
with a small "ms tcp read timeout", the timeout test always fails. in
this change, we use rados_watch3() which supports the "timeout" option,
so we can pass a timeout smaller than "ms tcp read timeout", and hence
the test can pass.

Fixes: http://tracker.ceph.com/issues/19312
Signed-off-by: Kefu Chai <kchai@redhat.com>