mgr: use handle_pyerror() instead of PyErr_Print()
If an exception is raised in any of the python module functions
invoked by mgr (for example, ImportError if rest_framework isn't
available when serve() in called in the rest module), you don't
actually see the exception anywhere.  PyErr_Print() allegedly
dumps to stderr, but that seems to evaporate into the aether.
If we replace this with calls to handle_pyerror() as in
PyModules::init(), exceptions are somewhat more helpfully
written to the mgr log.

This commit also fixes what looks to me like a missed call to
PyGILState_Release() in the failure case in MgrPyModule::serve().

Signed-off-by: Tim Serong <tserong@suse.com>