mgr: fix metadata handling from old MDS daemons
In abde684 we started always fetching metadata on fsmap changes
for legacy daemons (i.e. those without addr fields) -- this works
but is very inefficient.

By injecting an 'addr' field based on the address in
the FSMap, if the field is missing, we can reload
metadata only on daemon restarts.

Signed-off-by: John Spray <john.spray@redhat.com>