common/perf_counters: allow priorities to be assigned to perf counters
The idea is that higher priority counters can be surfaced by user tools
automatically (e.g., 'ceph daemonperf') without enshrining knowledge in
the tool of the internal workings of the daemon.  Similarly, the tool
may not know which counters are in use for a particular daemon (e.g.,
filestore vs bluestore).

This requires a bit of discipline in the ceph code to use consistent
priorities, but as long as developers are using the tools that consume
the priorities (and I expect they will if we make the counters useful)
we should converge on good values.

Signed-off-by: Sage Weil <sage@redhat.com>