os/bluestore: unify throttling model
Implement a super simple model for the cost of a transaction
for the purposes of throttling.  This replaces two independent
throttles (one for transaction ops, one for bytes) and puts
them both under the 'bytes' throttle.  The txc model cost is
expressed in terms of bytes as this is probably the simplest
thing for users to reason about.

Signed-off-by: Sage Weil <sage@redhat.com>