mgr/Mgr: fix map waits in init
If we go active multiple times we may already have an fsmap
subscription registered via MonClient, which means that we
won't receive a net new FSMap during the init phase.  This is
trivial to reproduce; just 'ceph mgr fail x' and when it
activates the second time init() will get stuck waiting.

The core issue is that monc subs are stateful, but we forget
the various maps with the previous Mgr incarnation.  Sipmle
workaround is to reset the mon connection as we init.

Signed-off-by: Sage Weil <sage@redhat.com>