msg/async/rdma: add rdma local configuration for systemd
issue: 992610

The script is only called if rdma is found in ceph.conf.

If ceph-conf or the script return an error, the comand/script is
ignored and won't affect the service, since both ExecStartPre and
ExecStopPost commands begin with the '-' sign.

Service is only reloaded if a change was made to any of the conf files.

No changes are kept after rdma is removed and services restarted (the
script restores previous configuration when stopping the service).

All changes are made in /run/systemd/system/ where all files are
considered temporary and are  only kept for the current run of the service.

Signed-off-by: DanielBar-On <danielbo@mellanox.com>
Change-Id: I57d30cc4a3a828282cd65ac2df4c8302e063c0dc

Conflicts:
	PendingReleaseNotes