also respond to lvm2 partitions coming as a "disk"
The lvm2 partitions don't come with the environtment variable 'DEVTYPE'
set to 'partition'; it instead comes with the variable set to 'disk'.
So this makeshift patch makes the udev rule less strict stopping it from
checking ENV{DEVTYPE}.
I believe this is a very naive approach and a preferable solution is
making the lvm2 partitions come with 'DEVTYPE' set correctly.

Signed-off-by: Gunwoo Gim (a.k.a. Nicho1as) <wind8702@gmail.com>