mon/OSDMonitor: update creating epoch if target osd changed
there is chance that the acting_primary of a PG being created changes
across osdmap updates, in that case we should update the new creating_pgs
with the mapping's epoch, not the stale one when the PG was originally
created. otherwise the new target OSD won't get the pg-create message.

Fixes: http://tracker.ceph.com/issues/19515
Signed-off-by: Kefu Chai <kchai@redhat.com>