pybind: store errno in Error exception also
also pass the errno and error string to Exception's __init__()
so we can print them all using str(err). otherwise str(err) will only
print out the errorstr, and the type information is lost. for example,
if ping_monitor() raises an exception, all we get is an error of:

  error calling ping_monitor

which is barely helpful. after this change:

  [Errno 22] error calling ping_monitor

is printed instead.

Signed-off-by: Kefu Chai <kchai@redhat.com>