os/bluestore/BlueFS: add write cache
Cache most recently written files in memory.  This is intended to
mitigate the impact of rocksdb's block cache invalidation that
happens as a result of compaction.  Recently compacted and written
sst files will stick around for a short period to prime the
rocksdb key and block caches.  This is normally something that the
operating system cache takes care of, but bluefs is intended to
operate without caching beneath it (in particular, when NVMEDevice
is used with SPDK, which has no cache).

Signed-off-by: Sage Weil <sage@redhat.com>