rgw: add the remove-x-delete feature to cancel swift object expiration
In openstack swift, it also support the feature to cancel the object expiration,
which could be found at last point in
https://docs.openstack.org/user-guide/cli-swift-set-object-expiration.html. we
can remove the object expiration by set 'X-Remove-Delete-At:'.

This patch also could fix the bug that when we set the object expiration and
then upload the same object to the container again. The previous object expiration
also works, which is not compatible with the openstack swift.

Fixes: http://tracker.ceph.com/issues/19074
Signed-off-by: Jing Wenjun <jingwenjun@cmss.chinamobile.com>
(cherry picked from commit 230429ebc4ac9b5840bb93c7e0eeb5edbb949106)

Conflicts:
	src/rgw/rgw_file.cc:
          - processor->complete() has no object_size argument
            (compression is not supported in Jewel)
	src/rgw/rgw_op.cc:
          - processor->complete() has no object_size argument
            (compression is not supported in Jewel)
          - RGWPostObj::execute(): the code changed a lot in master,
            adjust delete_at manually (patch(1) tries to apply
            the corresponding hunk to a different function)
	src/rgw/rgw_rest_swift.cc:
          - trivial: STREAM_IO in Jewel versus dump_header in master