osd: Do not subtract object overlaps from cache usage
PG group field num_bytes does not include space that is part of clone
object overlaps. Introduce a new field num_bytes_overlap that stands for
this space only, so that it can be used to strictly limit cache tier
usage with object storage backends that do not support storing overlaps
efficiently.

Signed-off-by: Michal Koutný <mkoutny@suse.com>